refactored saving function to plugins
