+++ src/app/components/Composer/plugins/paste-html.js
@@ -1,0 +1,22 @@
+// tools
+
+// return
+export function Paste(options) {
+	const { html } = options
+  return {
+  	onPaste(event, data, state) {
+			if (data.type !== "html") return
+			const { document } = html.deserialize(data.html)
+			return state
+				.transform()
+				.insertFragment(document)
+				.apply()
+    }
+  }
+}
+
+
+
+
+	
+
